Revert "Refactor packaging and update contributor guidance"
