+++ bin/topo_and_brdf_correction.py
@@ -1,14 +1,6 @@
 from argparse import ArgumentParser
 import json
-from pathlib import Path
-import sys
-
-PROJECT_ROOT = Path(__file__).resolve().parents[1]
-SRC_DIR = PROJECT_ROOT / "src"
-if str(SRC_DIR) not in sys.path:
-    sys.path.insert(0, str(SRC_DIR))
-
-from cross_sensor_cal.topo_and_brdf_correction import topo_and_brdf_correction
+from src.topo_and_brdf_correction import topo_and_brdf_correction
 
 if __name__ == "__main__":
     parser = ArgumentParser()
