--- pyproject.toml
@@ -1,73 +1,0 @@
-[build-system]
-requires = ["setuptools>=68", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "cross-sensor-cal"
-version = "0.1.0"
-description = "Cross-sensor calibration workflows for NEON hyperspectral data."
-readme = "README.md"
-requires-python = ">=3.9"
-license = { file = "LICENSE" }
-authors = [
-    { name = "Earth Lab" }
-]
-keywords = [
-    "remote-sensing",
-    "neon",
-    "spectroscopy",
-    "calibration",
-]
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3 :: Only",
-    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
-    "Operating System :: OS Independent",
-]
-dependencies = [
-    "geopandas",
-    "hytools",
-    "matplotlib",
-    "numpy",
-    "pandas",
-    "rasterio",
-    "requests",
-    "scipy",
-    "shapely",
-    "spectral",
-    "tqdm",
-]
-
-[project.optional-dependencies]
-dev = [
-    "black",
-    "mypy",
-    "pytest",
-    "ruff",
-]
-
-[project.urls]
-Homepage = "https://github.com/earthlab/cross-sensor-cal"
-Documentation = "https://github.com/earthlab/cross-sensor-cal/tree/main/docs"
-Issues = "https://github.com/earthlab/cross-sensor-cal/issues"
-
-[tool.setuptools.packages.find]
-where = ["src"]
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-addopts = "-ra"
-
-[tool.black]
-target-version = ["py310"]
-line-length = 88
-
-[tool.ruff]
-target-version = "py310"
-line-length = 88
-select = ["E", "F", "I", "B", "UP"]
-
-[tool.mypy]
-python_version = "3.10"
-ignore_missing_imports = true
-
