--- requirements-dev.txt
@@ -1,4 +1,0 @@
-black
-mypy
-pytest
-ruff
