+++ src/file_sort.py
@@ -2,7 +2,7 @@ import os
 import pandas as pd
 from pathlib import Path
 from typing import List, Tuple, Dict, Type
-from .file_types import (
+from src.file_types import (
     DataFile,
     NEONReflectanceFile,
     NEONReflectanceENVIFile,
