+++ src/standard_resample.py
@@ -11,11 +11,8 @@ import sys
 from scipy.interpolate import interp1d
 import glob
 
-from .file_types import (
-    NEONReflectanceBRDFCorrectedENVIHDRFile,
-    NEONReflectanceResampledHDRFile,
-    NEONReflectanceBRDFCorrectedENVIFile,
-)
+from src.file_types import NEONReflectanceBRDFCorrectedENVIHDRFile, NEONReflectanceResampledHDRFile, \
+    NEONReflectanceBRDFCorrectedENVIFile
 
 PROJ_DIR = os.path.dirname(os.path.dirname(__file__))
